refactor(settings): split out domains page
The drive settings General page had grown to own the Custom Domains card and its publishing controls alongside unrelated general settings (drive name, home page, members overview).

Extract the Custom Domains UI into a dedicated settings/domains route so each settings page has a single responsibility and General stays lean. Pure frontend move — same components, handlers, and API calls; no behavior change.

Register a 'Domains & Publishing' entry in the owner/admin settings navigation so the new route is reachable.
