+++ apps/web/src/app/dashboard/[driveId]/settings/general/page.tsx
@@ -5,20 +5,18 @@ import { useParams, useRouter } from 'next/navigation';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import { Badge } from '@/components/ui/badge';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { Skeleton } from '@/components/ui/skeleton';
-import { ChevronLeft, Shield, Users, Bot, Plug2, Loader2, Home, Globe, Trash2, Plus, RefreshCw, CheckCircle2, XCircle, Lock } from 'lucide-react';
+import { ChevronLeft, Shield, Users, Bot, Plug2, Loader2, Home } from 'lucide-react';
 import Link from 'next/link';
 import { useDriveStore } from '@/hooks/useDrive';
 import { usePageTree } from '@/hooks/usePageTree';
 import { findNodeAndParent } from '@/lib/tree/tree-utils';
 import { useEditingStore } from '@/stores/useEditingStore';
 import { toast } from 'sonner';
-import { fetchWithAuth, patch, del } from '@/lib/auth/auth-fetch';
+import { fetchWithAuth, patch } from '@/lib/auth/auth-fetch';
 import useSWR from 'swr';
-import { normalizeHostname, validateCustomDomain, buildDnsInstructions } from '@pagespace/lib/validators/custom-domain';
 
 interface DriveMember {
   id: string;
@@ -39,20 +37,6 @@ interface AppMembersResponse {
   appMembers: { id: string }[];
 }
 
-interface CustomDomain {
-  id: string;
-  driveId: string;
-  hostname: string;
-  status: 'pending' | 'verified' | 'failed' | 'dns_failed' | 'provisioning' | 'active' | 'cert_failed';
-  createdAt: string;
-}
-
-interface DomainsResponse {
-  domains: CustomDomain[];
-  /** Maximum custom domains allowed by the drive owner's plan (0 = not available). */
-  limit: number;
-}
-
 const fetcher = (url: string) => fetchWithAuth(url).then((r) => r.json());
 
 export default function GeneralSettingsPage() {
@@ -68,12 +52,6 @@ export default function GeneralSettingsPage() {
   const [name, setName] = useState('');
   const [isSaving, setIsSaving] = useState(false);
   const [isClearingHomePage, setIsClearingHomePage] = useState(false);
-  const [newDomain, setNewDomain] = useState('');
-  const [isAddingDomain, setIsAddingDomain] = useState(false);
-  const [removingDomainId, setRemovingDomainId] = useState<string | null>(null);
-  const [verifyingDomainId, setVerifyingDomainId] = useState<string | null>(null);
-  const [verifyReasons, setVerifyReasons] = useState<Record<string, string | undefined>>({});
-  const [refreshingCertId, setRefreshingCertId] = useState<string | null>(null);
   const startEditing = useEditingStore((s) => s.startEditing);
   const endEditing = useEditingStore((s) => s.endEditing);
 
@@ -104,10 +82,6 @@ export default function GeneralSettingsPage() {
     drive ? `/api/drives/${driveId}/apps/members` : null,
     fetcher
   );
-  const { data: domainsData, mutate: mutateDomains } = useSWR<DomainsResponse>(
-    drive && canManage ? `/api/drives/${driveId}/domains` : null,
-    fetcher
-  );
 
   const homePageId = drive?.homePageId ?? null;
   // Memoized: the controlled name input re-renders this page per keystroke,
@@ -116,114 +90,6 @@ export default function GeneralSettingsPage() {
     () => (homePageId ? findNodeAndParent(tree, homePageId)?.node ?? null : null),
     [tree, homePageId]
   );
-
-  const handleAddDomain = async () => {
-    const hostname = normalizeHostname(newDomain.trim());
-    const validation = validateCustomDomain(hostname);
-    if (!validation.valid) {
-      toast.error(validation.reason);
-      return;
-    }
-    if (isAddingDomain) return;
-    setIsAddingDomain(true);
-    try {
-      const res = await fetchWithAuth(`/api/drives/${driveId}/domains`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ hostname }),
-      });
-      if (!res.ok) {
-        const data = await res.json().catch(() => ({})) as { error?: string };
-        if (res.status === 409) {
-          toast.error('That domain is already registered');
-        } else if (res.status === 403) {
-          toast.error(data.error ?? 'Custom domains are not available on your current plan');
-        } else {
-          toast.error(data.error ?? 'Failed to add domain');
-        }
-        return;
-      }
-      setNewDomain('');
-      await mutateDomains();
-      toast.success('Domain added — set the DNS records below to activate it');
-    } catch {
-      toast.error('Failed to add domain');
-    } finally {
-      setIsAddingDomain(false);
-    }
-  };
-
-  const handleVerifyDomain = async (domainId: string) => {
-    if (verifyingDomainId) return;
-    setVerifyingDomainId(domainId);
-    try {
-      const res = await fetchWithAuth(`/api/drives/${driveId}/domains/${domainId}/verify`, {
-        method: 'POST',
-      });
-      const data = await res.json().catch(() => ({})) as { verified?: boolean; reason?: string; error?: string };
-      if (!res.ok) {
-        toast.error(data.error ?? 'Verification failed');
-        return;
-      }
-      setVerifyReasons((prev) => ({ ...prev, [domainId]: data.reason }));
-      await mutateDomains();
-      if (data.verified) {
-        toast.success('Domain verified — provisioning SSL cert…');
-        await handleRefreshCert(domainId);
-      } else {
-        toast.error(data.reason ?? 'DNS records not yet propagated');
-      }
-    } catch {
-      toast.error('Failed to verify domain');
-    } finally {
-      setVerifyingDomainId(null);
-    }
-  };
-
-  const handleRefreshCert = async (domainId: string) => {
-    if (refreshingCertId) return;
-    setRefreshingCertId(domainId);
-    try {
-      const res = await fetchWithAuth(`/api/drives/${driveId}/domains/${domainId}/cert/refresh`, {
-        method: 'POST',
-      });
-      const data = await res.json().catch(() => ({})) as { status?: string; error?: string };
-      if (!res.ok) {
-        if (res.status === 503) {
-          toast.error('SSL provisioning is not yet configured');
-        } else {
-          toast.error(data.error ?? 'Failed to refresh cert status');
-        }
-        return;
-      }
-      await mutateDomains();
-      if (data.status === 'active') {
-        toast.success('SSL certificate is active');
-      } else if (data.status === 'provisioning') {
-        toast.success('SSL cert provisioned — check back in a few minutes');
-      } else if (data.status === 'cert_failed') {
-        toast.error('SSL provisioning failed — click Retry SSL to try again');
-      }
-    } catch {
-      toast.error('Failed to refresh cert status');
-    } finally {
-      setRefreshingCertId(null);
-    }
-  };
-
-  const handleRemoveDomain = async (domainId: string) => {
-    if (removingDomainId) return;
-    setRemovingDomainId(domainId);
-    try {
-      await del(`/api/drives/${driveId}/domains/${domainId}`);
-      await mutateDomains();
-      toast.success('Domain removed');
-    } catch {
-      toast.error('Failed to remove domain');
-    } finally {
-      setRemovingDomainId(null);
-    }
-  };
 
   const handleClearHomePage = async () => {
     if (isClearingHomePage) return;
@@ -417,306 +283,6 @@ export default function GeneralSettingsPage() {
           </Button>
         </CardContent>
       </Card>
-
-      <CustomDomainsCard
-        domains={domainsData?.domains ?? []}
-        limit={domainsData?.limit ?? null}
-        newDomain={newDomain}
-        onNewDomainChange={setNewDomain}
-        onAdd={handleAddDomain}
-        onRemove={handleRemoveDomain}
-        onVerify={handleVerifyDomain}
-        onRefreshCert={handleRefreshCert}
-        isAdding={isAddingDomain}
-        removingId={removingDomainId}
-        verifyingId={verifyingDomainId}
-        refreshingCertId={refreshingCertId}
-        verifyReasons={verifyReasons}
-      />
-    </div>
-  );
-}
-
-// ── Custom Domains Card ───────────────────────────────────────────────────────
-
-interface CustomDomainsCardProps {
-  domains: CustomDomain[];
-  /** null = still loading; 0 = not available on plan; N = max allowed */
-  limit: number | null;
-  newDomain: string;
-  onNewDomainChange: (v: string) => void;
-  onAdd: () => void;
-  onRemove: (id: string) => void;
-  onVerify: (id: string) => void;
-  onRefreshCert: (id: string) => void;
-  isAdding: boolean;
-  removingId: string | null;
-  verifyingId: string | null;
-  refreshingCertId: string | null;
-  verifyReasons: Record<string, string | undefined>;
-}
-
-const EDGE_IPV4 = process.env.NEXT_PUBLIC_PUBLISH_EDGE_IPV4 ?? '';
-const EDGE_IPV6 = process.env.NEXT_PUBLIC_PUBLISH_EDGE_IPV6 ?? '';
-const CNAME_TARGET = process.env.NEXT_PUBLIC_PUBLISH_EDGE_CNAME_TARGET ?? '';
-
-function CustomDomainsCard({ domains, limit, newDomain, onNewDomainChange, onAdd, onRemove, onVerify, onRefreshCert, isAdding, removingId, verifyingId, refreshingCertId, verifyReasons }: CustomDomainsCardProps) {
-  const atCap = limit !== null && limit > 0 && domains.length >= limit;
-  const notAvailable = limit === 0;
-  const addDisabled = isAdding || !newDomain.trim() || atCap || notAvailable;
-
-  return (
-    <Card>
-      <CardHeader>
-        <div className="flex items-start justify-between">
-          <div>
-            <CardTitle className="flex items-center gap-2">
-              <Globe className="h-4 w-4" />
-              Custom Domains
-            </CardTitle>
-            <CardDescription>
-              Point your own domain at this drive&apos;s published canvas site
-            </CardDescription>
-          </div>
-          {limit !== null && limit > 0 && (
-            <span className="text-xs text-muted-foreground tabular-nums mt-1">
-              {domains.length} / {limit}
-            </span>
-          )}
-        </div>
-      </CardHeader>
-      <CardContent className="space-y-4">
-        {notAvailable ? (
-          <p className="text-sm text-muted-foreground">
-            Custom domains are not available on your current plan. Upgrade to Pro or higher to add a custom domain.
-          </p>
-        ) : (
-          <div className="flex gap-2">
-            <Label htmlFor="new-custom-domain" className="sr-only">Custom domain</Label>
-            <Input
-              id="new-custom-domain"
-              placeholder="e.g. docs.acme.com or acme.com"
-              value={newDomain}
-              onChange={(e) => onNewDomainChange(e.target.value)}
-              onKeyDown={(e) => e.key === 'Enter' && !addDisabled && onAdd()}
-              className="flex-1"
-              disabled={atCap}
-            />
-            <Button onClick={onAdd} disabled={addDisabled}>
-              {isAdding ? <Loader2 className="h-4 w-4 animate-spin" /> : <Plus className="h-4 w-4" />}
-              Add
-            </Button>
-          </div>
-        )}
-
-        {atCap && !notAvailable && (
-          <p className="text-xs text-muted-foreground">
-            Domain limit reached ({domains.length} / {limit}). Remove a domain to add another, or upgrade your plan.
-          </p>
-        )}
-
-        {domains.length > 0 ? (
-          <div className="space-y-3">
-            {domains.map((domain) => (
-              <DomainRow
-                key={domain.id}
-                domain={domain}
-                onRemove={onRemove}
-                onVerify={onVerify}
-                onRefreshCert={onRefreshCert}
-                isRemoving={removingId === domain.id}
-                isVerifying={verifyingId === domain.id}
-                isRefreshingCert={refreshingCertId === domain.id}
-                anyBusy={verifyingId !== null || refreshingCertId !== null}
-                verifyReason={verifyReasons[domain.id]}
-              />
-            ))}
-          </div>
-        ) : (
-          !notAvailable && <p className="text-sm text-muted-foreground">No custom domains yet</p>
-        )}
-      </CardContent>
-    </Card>
-  );
-}
-
-function DomainRow({
-  domain,
-  onRemove,
-  onVerify,
-  onRefreshCert,
-  isRemoving,
-  isVerifying,
-  isRefreshingCert,
-  anyBusy,
-  verifyReason,
-}: {
-  domain: CustomDomain;
-  onRemove: (id: string) => void;
-  onVerify: (id: string) => void;
-  onRefreshCert: (id: string) => void;
-  isRemoving: boolean;
-  isVerifying: boolean;
-  isRefreshingCert: boolean;
-  anyBusy: boolean;
-  verifyReason: string | undefined;
-}) {
-  const [showDns, setShowDns] = useState(false);
-  const instructions = buildDnsInstructions({
-    hostname: domain.hostname,
-    edgeIpv4: EDGE_IPV4,
-    edgeIpv6: EDGE_IPV6,
-    cnameTarget: CNAME_TARGET,
-  });
-
-  const statusBadge = () => {
-    if (domain.status === 'active') {
-      return (
-        <Badge variant="secondary" className="text-xs text-green-600 gap-1">
-          <Lock className="h-3 w-3" />
-          Active
-        </Badge>
-      );
-    }
-    if (domain.status === 'provisioning') {
-      return (
-        <Badge variant="secondary" className="text-xs gap-1">
-          <Loader2 className="h-3 w-3 animate-spin" />
-          Provisioning SSL
-        </Badge>
-      );
-    }
-    if (domain.status === 'verified') {
-      return (
-        <Badge variant="secondary" className="text-xs text-green-600 gap-1">
-          <CheckCircle2 className="h-3 w-3" />
-          DNS Verified
-        </Badge>
-      );
-    }
-    if (domain.status === 'dns_failed') {
-      return (
-        <Badge variant="secondary" className="text-xs text-destructive gap-1">
-          <XCircle className="h-3 w-3" />
-          DNS Failed
-        </Badge>
-      );
-    }
-    if (domain.status === 'cert_failed') {
-      return (
-        <Badge variant="secondary" className="text-xs text-orange-600 gap-1">
-          <XCircle className="h-3 w-3" />
-          SSL Failed
-        </Badge>
-      );
-    }
-    if (domain.status === 'failed') {
-      return (
-        <Badge variant="secondary" className="text-xs text-destructive gap-1">
-          <XCircle className="h-3 w-3" />
-          DNS Failed
-        </Badge>
-      );
-    }
-    return (
-      <Badge variant="secondary" className="text-xs">
-        Pending DNS
-      </Badge>
-    );
-  };
-
-  const showVerifyButton = domain.status === 'pending' || domain.status === 'failed' || domain.status === 'dns_failed';
-  const showCertButton = domain.status === 'verified' || domain.status === 'provisioning' || domain.status === 'cert_failed';
-
-  return (
-    <div className="border rounded-md p-3 space-y-2">
-      <div className="flex items-center justify-between gap-2">
-        <div className="flex items-center gap-2 min-w-0">
-          <Globe className="h-4 w-4 text-muted-foreground flex-shrink-0" />
-          <span className="text-sm font-medium truncate">{domain.hostname}</span>
-          {statusBadge()}
-        </div>
-        <div className="flex items-center gap-1 flex-shrink-0">
-          {showVerifyButton && (
-            <Button
-              variant="outline"
-              size="sm"
-              className="text-xs h-7"
-              onClick={() => onVerify(domain.id)}
-              disabled={anyBusy}
-              aria-label={`Verify domain ${domain.hostname}`}
-            >
-              {isVerifying ? (
-                <Loader2 className="h-3 w-3 animate-spin mr-1" />
-              ) : (
-                <RefreshCw className="h-3 w-3 mr-1" />
-              )}
-              {(domain.status === 'failed' || domain.status === 'dns_failed') ? 'Re-check DNS' : 'Verify DNS'}
-            </Button>
-          )}
-          {showCertButton && (
-            <Button
-              variant="outline"
-              size="sm"
-              className="text-xs h-7"
-              onClick={() => onRefreshCert(domain.id)}
-              disabled={anyBusy}
-              aria-label={`Check SSL cert for ${domain.hostname}`}
-            >
-              {isRefreshingCert ? (
-                <Loader2 className="h-3 w-3 animate-spin mr-1" />
-              ) : (
-                <Lock className="h-3 w-3 mr-1" />
-              )}
-              {domain.status === 'provisioning' ? 'Check SSL' : domain.status === 'cert_failed' ? 'Retry SSL' : 'Provision SSL'}
-            </Button>
-          )}
-          <Button
-            variant="ghost"
-            size="sm"
-            className="text-xs h-7"
-            onClick={() => setShowDns((p) => !p)}
-          >
-            {showDns ? 'Hide DNS' : 'Show DNS'}
-          </Button>
-          <Button
-            variant="ghost"
-            size="sm"
-            className="h-7 w-7 p-0 text-destructive hover:text-destructive"
-            onClick={() => onRemove(domain.id)}
-            disabled={isRemoving}
-            aria-label={`Remove domain ${domain.hostname}`}
-          >
-            {isRemoving ? <Loader2 className="h-3 w-3 animate-spin" /> : <Trash2 className="h-3 w-3" />}
-          </Button>
-        </div>
-      </div>
-
-      {(domain.status === 'failed' || domain.status === 'dns_failed') && verifyReason && (
-        <p className="text-xs text-destructive bg-destructive/5 rounded px-2 py-1">{verifyReason}</p>
-      )}
-
-      {showDns && (
-        <div className="bg-muted rounded p-3 space-y-2">
-          <p className="text-xs text-muted-foreground">
-            {instructions.isApex
-              ? 'Add these records at your DNS provider:'
-              : 'Add this record at your DNS provider:'}
-          </p>
-          <div className="space-y-1">
-            {instructions.records.map((r, i) => (
-              <div key={i} className="flex gap-3 text-xs font-mono">
-                <span className="w-12 text-muted-foreground">{r.type}</span>
-                <span className="w-8 text-muted-foreground">{r.name}</span>
-                <span className="break-all">{r.value}</span>
-              </div>
-            ))}
-          </div>
-          <p className="text-xs text-muted-foreground mt-2">
-            Once records propagate, click Verify DNS to confirm setup and automatically provision SSL.
-          </p>
-        </div>
-      )}
     </div>
   );
 }
