+++ apps/web/src/app/dashboard/[driveId]/settings/page.tsx
@@ -5,7 +5,7 @@ import { useEffect } from 'react';
 import { useParams, useRouter } from 'next/navigation';
 import { Button } from '@/components/ui/button';
 import { Skeleton } from '@/components/ui/skeleton';
-import { ChevronLeft, Shield, Users, Brain, Cable, HardDrive, Trash2, SlashSquare } from 'lucide-react';
+import { ChevronLeft, Shield, Users, Brain, Cable, HardDrive, Trash2, SlashSquare, Globe } from 'lucide-react';
 import { useDriveStore } from '@/hooks/useDrive';
 import { SettingsRow, type SettingsItem } from '@/app/settings/SettingsRow';
 
@@ -69,6 +69,13 @@ export default function DriveSettingsPage() {
               description: 'Drive name and member overview',
               icon: Shield,
               href: `/dashboard/${driveId}/settings/general`,
+              available: true,
+            },
+            {
+              title: 'Domains & Publishing',
+              description: 'Custom domains for your published site',
+              icon: Globe,
+              href: `/dashboard/${driveId}/settings/domains`,
               available: true,
             },
             {
